# features: ## throughput of all hosts ## set throughput of all hosts ## priority of hosts ## changing tcp CC of hosts

+++ admin/src/main/java/org/admin/backend/service/dtos/request/HostRequest.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class HostRequest {
   private String ip;
-  private Integer port;
-    private Boolean isActive;
+  private Long port;
+  private Boolean isActive;
 }
