+++ admin/src/main/java/org/admin/backend/service/dtos/response/HostResponse.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class HostResponse {
   private String ip;
-  private Integer port;
+  private Long port;
   private Boolean isActive;
 }
